s2491 - Modify opt3001 driver for WBLGT01
WBLGT01 does not have internal OPT3001 so an external OPT3001 sensor is used (MikroElecktronica Ambient 2 click board).
This external sensor uses 3.3V but on WBLGT01, zephyr drivers initialization is run at boot with 1.8V and OPT3001 external sensor initialization can not be done (because level shifter power is not enabled).
The flag "status" in dts file is used to make different driver initializations:
if the flag value is "reserved", zephyr drivers initialisation for OPT3001 just binds the address. The driver initialization should be done in external Zephyr code.
If the flag value is "okay", zephyr driver initialization is done normally.

Change-Id: I0746236deedf2c90998fa1b848ab74e74686061b
Signed-off-by: Nicolas <nicolas.pelissier@wiifor.com>